Move AddHosting to DI namespace
