+++ src/Microsoft.AspNet.Hosting/HostingServices.cs
@@ -4,9 +4,6 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Microsoft.AspNet.Hosting.Builder;
-using Microsoft.AspNet.Hosting.Server;
-using Microsoft.AspNet.Hosting.Startup;
 using Microsoft.Framework.ConfigurationModel;
 using Microsoft.Framework.DependencyInjection;
 using Microsoft.Framework.DependencyInjection.ServiceLookup;
@@ -39,42 +36,6 @@ namespace Microsoft.AspNet.Hosting
             var services = Import(fallbackServices);
             services.AddHosting(configuration);
             services.AddSingleton<IServiceManifest>(sp => new HostingManifest(fallbackServices));
-            services.AddInstance<IConfigureHostingEnvironment>(new ConfigureHostingEnvironment(configuration));
-            return services;
-        }
-
-        // REVIEW: Logging doesn't depend on DI, where should this live?
-        public static IServiceCollection AddLogging(this IServiceCollection services, IConfiguration config = null)
-        {
-            var describe = new ServiceDescriber(config);
-            services.TryAdd(describe.Singleton<ILoggerFactory, LoggerFactory>());
-            return services;
-        }
-
-        public static IServiceCollection AddHosting(this IServiceCollection services, IConfiguration configuration = null)
-        {
-            var describer = new ServiceDescriber(configuration);
-
-            services.TryAdd(describer.Transient<IHostingEngine, HostingEngine>());
-            services.TryAdd(describer.Transient<IServerManager, ServerManager>());
-
-            services.TryAdd(describer.Transient<IStartupManager, StartupManager>());
-            services.TryAdd(describer.Transient<IStartupLoaderProvider, StartupLoaderProvider>());
-
-            services.TryAdd(describer.Transient<IApplicationBuilderFactory, ApplicationBuilderFactory>());
-            services.TryAdd(describer.Transient<IHttpContextFactory, HttpContextFactory>());
-
-            services.TryAdd(describer.Instance<IApplicationLifetime>(new ApplicationLifetime()));
-
-            services.AddTypeActivator(configuration);
-            // TODO: Do we expect this to be provide by the runtime eventually?
-            services.AddLogging(configuration);
-            // REVIEW: okay to use existing hosting environment/httpcontext if specified?
-            services.TryAdd(describer.Singleton<IHostingEnvironment, HostingEnvironment>());
-
-            // TODO: Remove this once we have IHttpContextAccessor
-            services.AddContextAccessor(configuration);
-
             return services;
         }
